Fix travis builds
LINUX: old gcc need explicit types for enum types for hasher
ANDROID: glm::max does not work on float types :(. But probably better use the internal MAX macro we
have

+++ core/src/style/pointStyle.h
@@ -87,7 +87,7 @@ namespace std {
             hash_combine(seed, p.color);
             hash_combine(seed, p.size.x);
             hash_combine(seed, p.size.y);
-            hash_combine(seed, p.placement);
+            hash_combine(seed, (int)p.placement);
             hash_combine(seed, optionsHash(p.labelOptions));
             return seed;
         }
